Inspection/requirement.txt
asgiref==3.4.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
charset-normalizer==2.0.7
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==35.0.0
daphne==3.0.2
defusedxml==0.7.1
Django==3.2.9
django-bitfield==2.1.0
django-cors-headers==3.10.0
django-js-asset==1.2.2
django-map-widgets==0.3.2
django-mptt==0.13.4
django-separatedvaluesfield==0.6.0
django-templated-mail==1.1.1
django-timezone-field==4.2.1
django-uuidfield==0.5.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
djoser==2.1.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
itypes==1.2.0
Jinja2==3.0.3
lxml==4.6.4
MarkupSafe==2.0.1
oauthlib==3.1.1
Pillow==8.4.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.3.0
pyOpenSSL==21.0.0
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
service-identity==21.1.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
South==1.0.2
sqlparse==0.4.2
Twisted==21.7.0
twisted-iocpsupport==1.0.2
txaio==21.2.1
typing-extensions==3.10.0.2
unicodecsv==0.14.1
uritemplate==4.1.1
urllib3==1.26.7
zope.interface==5.4.0
